+++ nengo/tests/test_node.py
@@ -379,3 +379,31 @@ def test_node_with_unusual_strided_view(Simulator, seed):
 
     with Simulator(model):
         pass
+
+
+def test_non_finite_values(Simulator):
+    with nengo.Network() as model:
+        with pytest.raises(ValidationError):
+            node = nengo.Node(np.inf)
+
+    with nengo.Network() as model:
+        with pytest.raises(ValidationError):
+            node = nengo.Node(np.nan)
+
+    with nengo.Network() as model:
+        node = nengo.Node(lambda t: np.inf)
+        ens = nengo.Ensemble(10, 1)
+        nengo.Connection(node, ens)
+
+    with Simulator(model) as sim:
+        with pytest.raises(SimulationError):
+            sim.run(0.01)
+
+    with nengo.Network() as model:
+        node = nengo.Node(lambda t: np.nan)
+        ens = nengo.Ensemble(10, 1)
+        nengo.Connection(node, ens)
+
+    with Simulator(model) as sim:
+        with pytest.raises(SimulationError):
+            sim.run(0.01)